Reject old CLI authentication
Summary: Old CLI will return 200 even if you use the wrong HTTP method. I thought it was failing so I went ahead with the above. However, users would login with old auth and get a success message in the UI but a failure in the CLI. his change will now return non-200 on failures and will return an 'OK' on success. The UI now looks for the 'OK' http body to know that the auth worked properly.

Test Plan: tested with old cli -> auth no longer works and the user is notified both in the CLI and the UI. New CLI, they can log in.

Reviewers: michelle, vihang, nlanam, #engineering

Reviewed By: vihang, #engineering

JIRA Issues: PC-769

Differential Revision: https://phab.corp.pixielabs.ai/D7371

GitOrigin-RevId: 2c11697f74e0b0e29c96acdb3fb1d4f89e1be88e