Actually fix cloud release job
Summary:
for some reason, the staging build I'd run to test this had passed, so I thought it worked. it obviously did not work, after trying to run the prod build yesterday.
this is because the prerelease part of the tag shouldn't contain any non-ASCII characters. before I was replacing the "-"'s with "/"s. instead, we should just remove the invalid characters.

Test Plan: ran staging release

Reviewers: vihang, zasgar, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7376

GitOrigin-RevId: d16211d3d8901437517b3c786b3996b7749c3a56