Fix k8s watcher resource versions
Summary: when we start the k8s watcher, we need to specify the resource version at which you should start watching at. if you specify one thats too old, then the watcher will complain and error out. to get the resource version, I was assuming that the resource list that we were getting from the k8s API was ordered and taking the resource version of the last item in the list. that wasnt necessarily true, and things were hitting the "k8s resource version too old" error. instead, you are supposed to get the resourceVersion from the returned list object.

Test Plan: ran metadataIntegration branch in skaffold

Reviewers: vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7400

GitOrigin-RevId: 44d964f9d1fdb7276a0753ed12b83be55244cd14