Fix bug where md handler doesnt correctly send container updates for deleted pods
Summary:
while integrating everything together, I ran into a bug where mds would fail to send some container updates to the agent. this is because the PodProcessor's GetUpdatesToSend relies on the PodIP map to find which IPs the container updates should be sent to. however, the entry for that pod is cleared from the podIP map if that pod is deleted. this is what was happening:
1. update for terminated pod with some containers is sent to the handler
2. the handler clears out the pod from the podIP map, since it has terminated.
3. the handler tries to use the podIP map to get thet IP which the container update should be sent to. it cant find the podIP, so doesnt send the container update to the agent.
instead, the GetUpdatesToSend function should use the PodUpdate included in the args to get the hostIP. this PodUpdate should always be in the args.

Test Plan: ran in skaffold, update unit test

Reviewers: vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7397

GitOrigin-RevId: daada25e57c60d20a8462ecbe3f9a0562797ca0b