PP-2375: Heartbeats should only send out a bloom filter when the bloom filter has changed
Summary: We send out a lot of metadata bloom filters, even when the bloom filter contents have not changed. We previously tried to not send redundant data by checking the epoch ID on the metadata state. However, that is incremented every 5 seconds, even if the metadata filter itself has had no changes. As a result, I added a metadata filter epoch ID so that we can better discern if there has been a change in the filter. Also, we need to make sure we still send the metadata filter (Even if it hasn't changed) if the agent reregisters with the metadata service, so that logic should work as well.

Test Plan: I updated the tests and also refactored them to use the fake agent state manager, because we were previously testing a lot of the logic within the real agent state manager. I also extended our test coverage in general for heartbeats.

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: vihang, #engineering

JIRA Issues: PP-2375

Differential Revision: https://phab.corp.pixielabs.ai/D7382

GitOrigin-RevId: 845fff7504e1aa4bc4236a8ab094394f05d5d06c