PX-99 Cleanup YAML generator used in CLI deploy
Summary:
we want to improve our Helm support, to accomodate for customers that use Helm (such as customer). our only existing Helm chart hasnt been updated in a long time, and we rely on our update process to get the updated YAMLs.
However, as we move away from our upgrade job towards a scheme where all of our resources are present in the extracted YAML, it's important we routinely keep our Helm charts up to date and start adding it to our release builds. we should also allow more configurability in the helm charts--only deployKey is configurable at the moment.

this initial diff is just some cleanup. we originally used the YAMLGenerator to generate the deployed YAMLs and fill in the template. This diff generalizes the YAMLGenerator extract/executeTemplate into utils, that have no dependency on the actual Vizier yamls.

Test Plan: deployed through CLI with/without --extract_yamls

Reviewers: zasgar, vihang, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7416

GitOrigin-RevId: bc1d072054816fa4b30b015b5cc4b7bd1a6b6ba7