Correct copy-paste oversight; more robust env var replacement in Webpack
Summary: In D7337, `BUILD_TIMESTAMP` was changed to `STABLE_BUILD_TIMESTAMP` in `webpack.config.js` like the other vars above it. However, that specific one wasn't supposed to be renamed. This fixes that oversight, and also addresses a newfound issue (webpack version update?) where `EnvPlugin` has a bit of trouble with object destructuring.

Test Plan: Attempt to run locally. It should work again. Before this change, an error was thrown at runtime because `process.env` was left in the code instead of being replaced.

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: vihang, michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7432

GitOrigin-RevId: a6d3724a4bfd1208778e8789798b497b9d5a2221