Fix MDS bug where agent misses resources that havent updated for >24h
Summary:
we currently have a bug because we expire updates that are older than 24 hours. if there is a long-running resource that hasnt updated for over 24h, for example namespaces, if an agent is restarted, it will not get the update for that resource.
our solution is to send "refresh" updates every 30mins, so that even if an agent restarts, it will always be able to get info about all running resources since the last 24 hours.
since the login for fetching current resources is what the "sync" does, I put that logic in a loop.

Test Plan: ran with a bunch of logs, to make sure things flowed as expected

Reviewers: vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7433

GitOrigin-RevId: 09f73ddbd2a08458737861ede1d6eb7ad1a60143