PX-101 Add online boutique to px demo
Summary:
To assist with the pixienaut demo and tutorial for redis on Thursday, we want to make it easier to deploy online boutique.
This diff makes it possible with px demo. The source code for online boutique has its own set of quirks compared to the quirks of sock-shop,
so some of their handling is different.

Also this removes the concept of a "staging" bucket and replaces it with a "dev" bucket for develpers at Pixie who want to develop a new
demo application. Previously, prod access was required to even test something out becasue prod buckets were written to. Now, there is a dev
option to write to a dev bucket that Pixie developers (as in employees, not community) have access to.

Test Plan: ran the demo app, loaded it in the browser

Reviewers: michelle, zasgar, #engineering, vihang

Reviewed By: michelle, #engineering, vihang

Subscribers: vihang, htroisi

JIRA Issues: PX-101

Differential Revision: https://phab.corp.pixielabs.ai/D7437

GitOrigin-RevId: 5875102f994cff46c4aa95c38e75f9726a34e253