Correct an enum in schema.graphql; copy it to @pixie/api
Summary:
VZ_ST_ is the internal prefix for cluster status. The external one, and the one that is actually used by the GraphQL API, uses CS_ instead.
This change then updates the `update.sh` script to reliably generate the updated `schema.d.ts` file as well as copy it where `@pixie/api` can find and reexport it.

Test Plan: N/A (should be a no-op until the newly exported symbols see use)

Reviewers: michelle, vihang, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7452

GitOrigin-RevId: 033f4a9bb176f18a63b2ec7eef7014c130880db5