PP-2461: Make agent metadata state more resilient to missing updates.
Summary: We are currently not resilient to missing updates, because if a resource update points to another resource that we haven't heard about, then we LOG(FATAL). It's possible that we will lose some updates and our system needs to be resilient to that. Instead of fatal we LOG(INFO) now for this case.

Test Plan: We had very few tests for the metadata state so I added some.

Reviewers: michelle, vihang, vjain, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2461

Differential Revision: https://phab.corp.pixielabs.ai/D7454

GitOrigin-RevId: 985b404283f876eb20925fcbc3dadc4d374efc1c