Use `@pixie/api-react` in `@pixie/ui` code, part 1
Summary:
Following D7663 (and several predecessors to resolve blockers), this change uses the new `@pixie/api-react` hooks and contexts in the main UI code.
There are still some leaky abstractions and incomplete refactors to address after this, but it's a stable checkpoint to work with.

Big changes that still need to happen (later diffs):
* `VizierGRPCClientContext`, `Script[s]Context`, and some others need to move most of their logic out to `@pixie/api` and `@pixie/api-react`, respectively. Doing this will allow dropping a couple of nasty workarounds and cleaning up dependencies.
* `@pixie/api` is mostly unimplemented, and `@pixie/api-react` is largely avoiding gRPC work at the moment.
* `@pixie/ui` will need to become its own package in `packages` rather than a root above the others. This happens later and is mostly a cleanliness/configuration improvement, not urgent. Once this is one, we can introduce Lerna to simplify and decouple configuration even further.

Test Plan: Smoke test - this affects nearly every page and numerous actions. Particularly running scripts in the live UI, using the profile menu, and create/read/delete on API and Deployment keys. All behavior should appear unchanged, as with most of the ever-growing chain of tech-debt-removing and open-source-preparing changes.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-738

Differential Revision: https://phab.corp.pixielabs.ai/D7583

GitOrigin-RevId: 41efb8c37b5da6cd3b562b80dfbd231a0006d962