PX-99 Add util for templatizing k8s yamls
Summary:
this is part of the work to add better configurability to our helm charts and add them to our release build. currently, the only templateable field is "deployKey", which is hardcoded into our YAMLs.
we want to move towards a direction where all of the fields configurable via the CLI (annotations, labels, etcd, etc) are also configurable through helm. helm does this through templates, which means that our YAMLs should be templates.
the most common way of doing this is to just check the template YAMLs into the repo. however, we want our YAMLs still be deployable through skaffold and work with kustomize, as it is in its current state. however, keeping both the template YAML + deployable YAML checked into our repo can lead to consistency problems.

instead the proposed scheme is:
- keep our YAMLs as is, as deployable YAMLs
- add a util for making these YAMLs templateable (this diff)
- have the CLI write the templated files to the helm chart for a release build
- when deploying through the CLI, the CLI generates the templated files and fills in the template with the values passed in through the CLI (already implemented)

this diff adds the util for making YAMLs templates. unfortunately, we can't directly modify the YAML to insert the templates--as adding in the templates would make the YAML invalid. instead, we first have to:
1.  update the YAML to insert a placeholder in the correct place in the YAML. This uses k8s strategic patch. At this point, it is still a valid YAML.
2. this placeholder is then replaced with the actual template value. this is done by a string replace.
one of the hackier parts with (2), is that if the template value is a multiline string, the multiline string must have the correct identations. suggestions here are very welcome.

this diff uses the util to add DeployKey and Annotations, as examples for how it should work. future diffs will make all other configurable fields templates.
example template yaml:
```
apiVersion: v1
kind: Secret
metadata:
  annotations:
    {{range $element := split "," .Values.CustomAnnotations -}}
    {{ $kv := split "=" $element -}}
    {{if eq (len $kv) 2 -}}
    {{ index $kv 0}}: "{{ index $kv 1}}"
    {{- end}}
    {{end}}
  creationTimestamp: null
  labels:
    hello: test
  name: pl-deploy-secrets
  namespace: pl
stringData:
  deploy-key: {{ .Values.DeployKey | required "...." }}
```

Test Plan: extracted YAMLs, ensured it looked as expected, and deployed.

Reviewers: zasgar, vihang, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7447

GitOrigin-RevId: e99933dd1c376c0abba5f74002b1fd98c71294f2