PP-2458: Remove metadata stats handler
Summary: We used this to collect metadata stats in order to do the metadata redesign, but now we don't need it anymore and the refactor made it no longer work. It's not worth the effort of porting it over because it served its purpose, so it's removed.

Test Plan: n/a

Reviewers: vihang, michelle, vjain, #engineering

Reviewed By: vihang, #engineering

JIRA Issues: PP-2458

Differential Revision: https://phab.corp.pixielabs.ai/D7467

GitOrigin-RevId: bf47fc7be83560bd3b116a8eb859e1eec63b5e83