PP-2373: Agents shouldn't die on heartbeat timeout
Summary: We want to make the agents and the query broker (Vizier components that depend on the metadata service) more resilient to failures in the metadata service. Currently, if the metadata service is having a problem, agents will kill themselves again and again until the problem is resolved. This causes unnecessary churn on the system. Instead, the agents should quietly collect their data and wait for the problem to resolve.

Test Plan: updated

Reviewers: vihang, michelle, zasgar, vjain, #engineering

Reviewed By: vihang, #engineering

JIRA Issues: PP-2373

Differential Revision: https://phab.corp.pixielabs.ai/D7468

GitOrigin-RevId: 7c43b45388907d736f4128fef0897de4b37e0fd0