Add select UDF
Summary:
Adds the ability to select a value based on a boolean condition. The main issues with this is that it does eager evaluation of both sides.

We can reconsider how to fix this in the future, but current approches we use for this functionality are even more inefficient.

Test Plan: Added unit tests.

Reviewers: michelle, vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7476

GitOrigin-RevId: cf63188dfb6155d1765478265dd8893af22750d4