PX-99 Make all configurable fields in deploy yamls templates
Summary:
this is part of the work to add better configurability to our helm charts and add them to our release build.
D7447 added the util for adding templates to our deploy yamls, and updated the deployKey and customAnnotations field as an example. in this diff, the rest of the configurable fields are updated to be templates. now, we can finally configure all of these values in helm as well.
also updated the funcMap for the golang templates to use Sprig, since that is what helm uses.

followup diffs will actually add helm charts to our release build.

Test Plan: deployed with/without extract yamls. also extracted the templated yamls and tried them out in helm, to verify they actually work with helm.

Reviewers: vihang, zasgar, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7469

GitOrigin-RevId: 7e05882d1094df4cf899bddd6c87bcf84ad760d2