Shortcut eslint version check
Summary:
We run `arc lint` on every jenkins run, which causes every linter
to trigger `getVersion`. Since we run `eslint` with this helper script which
calls `yarn install`, this means we end up spending a lot of time installing
node modules even on diffs that might not trigger eslint.

This adds a `-v` opt to our script that short-circuits the install and returns
the currently installed eslint version.

This should speed up jenkins runs for non UI diffs.

Test Plan: N/A

Reviewers: zasgar, nlanam, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7483

GitOrigin-RevId: 71f51b05db97f72e320b730ae2800057eb2d0d48