Add px generate-yamls for generating helm charts
Summary:
this is a hidden command, similar to `px create-bundle`.
the command just generates the templated YAML files for helm, and writes it to a tar. this will be used by the jenkins release build. initially i was going to add it to `px deploy`, but there are already enough things in that command as is.

Test Plan: ran `px generate-yamls` and `px deploy` to make sure they both work as expected.

Reviewers: vihang, zasgar, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7485

GitOrigin-RevId: 57a8a23a33cb127d0da14378795227754a967d11