PP-2450: When the result plan is empty for a query, query broker should return an error, don't just silently pretend to launch it
Summary: We had a problem where Kelvins were being disconnected and there would be no query plan coming out of the heartbeats (because the query plan needs to have a Kelvin). So the query broker would say it isn't getting healthy results from the health check, but none of the agents ever received a query in the first place because the plan was empty. We should surface this case rather than silently ignoring the fact that there is no query being run at all.

Test Plan: added a unit test

Reviewers: vihang, michelle, vjain, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2450

Differential Revision: https://phab.corp.pixielabs.ai/D7490

GitOrigin-RevId: 2cb26dfbc5d678676b01d2df8056d7196737c650