Show an asterisk for required arguments.
Summary:
This is a stopgap for a more serious design update for argument handling in the UI. Arguments can be made required by omitting them in the Vis spec, as a reminder:
- If `defaultValue` is omitted, the user *must* provide a value themselves.
- If `defaultValue` is the empty string, the argument is treated as optional. The user does not need to provide a value.
- If `defaultValue` is a non-empty string, the argument will take on that value if the user does not provide one themselves.
This change does not enforce required arguments beyond the error that already occurs on the backend. That will come later with the more serious efforts.

Test Plan: select a script that has arguments. Make one of them required by removing its `defaultValue` property (as of this writing, no script in the `px/` namespace has such an argument). Run the script so that the UI picks up the change. An asterisk should appear next to that argument's name.

Reviewers: michelle, vihang, #engineering

Reviewed By: vihang, #engineering

JIRA Issues: PC-548

Differential Revision: https://phab.corp.pixielabs.ai/D7493

GitOrigin-RevId: 11625693553f80a8d990be97ecfd9c1f0439fd4a