Fix CLI to use current cluster by default. Add y/n check to px update to match px deploy.
Summary: px debug log, px live, and px update were still using the first healthy vizier. this could actually cause an update for px update, because it checks the first healthy vizier to see if the update succeeded, and if another vizier was deployed at that time, it could mix the two up. Also, we want parity in our CLI about what is the default behavior when -c is passed. While I was looking at this, I noticed that we should also add a manual y/n to the user when they update a cluster, similar to deploy.

Test Plan: ran these commands via bazel

Reviewers: michelle, zasgar, vihang, vjain, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7495

GitOrigin-RevId: 12ebfca8534b6dec763ec7f56c144bd398390dcb