TaskStructOffsetsResolver: run as subprocess to support threaded environments.
Summary:
For the resolver to work, the group_leader pointer must point to itself. This is only true for the main thread.

Unfortunately, that means if Stirling is not the main thread, it will fail.

This changes TaskStructOffsetsResolver to run inside a SubProcess, in which case it will be the main thread.
Results are sent back via a pipe.

Test Plan: Tests added

Reviewers: yzhao, jps, #engineering

Reviewed By: jps

Differential Revision: https://phab.corp.pixielabs.ai/D7479

GitOrigin-RevId: 89ad20941f72dc8528ed4426de93d281fac6f363