PP-2472 Use autoUpdateEnabled flag in Pixie Cloud to stop autoupdates
Summary: previous diffs added the flag to the backend, and added it to the CLI. this diff actually uses the flag to determine whether or not a cluster should update

Test Plan: ran dev cloud. deployed cluster with autoUpdateEnabled false, with an older vizier version. confirmed it stayed at the old version. reenabled the flag, and saw the cluster automatically update.

Reviewers: vihang, zasgar, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7506

GitOrigin-RevId: 37323b7b599629a515a1b02e8c17eccc9e8dc01f