Move CLI cluster type checks to allowlist model
Summary:
we are seeing cases where users are deploying to clusters that we don't support.
following the discussion on Slack, we want to move to more of a whitelist/allowlist model. for clusters that we know we do support, we do not display any errors/warnings. for clusters that we aret sure if we support (such as custom k8s clusters), then we still allow them to deploy, but notify them that things may not work as expected.
i decided to still keep some of the blacklist model, since we have cluster types that we know we do not support. we might as well prevent the deployment in these cases.

feedback on this flow, and the copy of the messaging, are appreciated.

Test Plan: ran deploy on minikube/gke.

Reviewers: oazizi, zasgar, philkuz, #engineering

Reviewed By: oazizi, zasgar, #engineering

Subscribers: htroisi

Differential Revision: https://phab.corp.pixielabs.ai/D6949

GitOrigin-RevId: 16873f84d6cc95a210db14513a07cd0a832f8dd9