PX-100 Allow users to specify cluster name when deploying
Summary:
before we wouldnt allow users to specify cluster name during deploys. if we were able to get the cluster name from the kubeconfig, we used that. otherwise, we would assign a randomly generated name.
this diff updates the CLI so that they can specify the cluster name. if none is specified, we will try the kubeconfig, then a random name.
it also updates the cloud so that if they specify a cluster name for a cluster with an existing name, we overwrite that name.

Test Plan: ran dev cloud, unit tests

Reviewers: zasgar, vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7514

GitOrigin-RevId: b7e615c3be2b11f76305aa6c5ddf8f765ac0d17f