PP-2445: Add source executor to scalar UDF (similar to UDTF)
Summary: Similar to UDTFs, some UDFs should only be executed on Kelvin or PEM Carnot instances. This diff follows the same protocol used for UDTF in order to add that distinction for UDF. This will be used to help plan queries so that Kelvin-only UDFs are scheduled on Kelvin, and also alert the user with a compiler error if they are trying to run a PEM-only UDF after a blocking node (aka on a Kelvin node). While some aspects of how we handle metadata and compilation of distributed plans will change in the future with Carnot ETL, the concept of certain UDFs needing to execute in on a specific type of agent is expected to stick around in some form.

Test Plan: updated unit tests

Reviewers: philkuz, zasgar, jamesbartlett, michelle, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D7528

GitOrigin-RevId: 2e8674fd914978926bb3f0e52831f175acb76bcb