bazel_build_deps.sh: Include libraries too
Summary:
Makes sure that if a source file affects a library but not a binary, the dependent library is still added to the buildables.

Triggered by a case where a file was under development, with no enabled test, which ended up breaking a coverage build on main.

Test Plan: Jenkins

Reviewers: vihang, zasgar, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7507

GitOrigin-RevId: 032de79b3e3d550bd9ab7fff2db8e7ba1db7bb71