Prevent running a script without specifying required args; warn
Summary:
An argument to a script is required if its variable definition omits `defaultValue`. The UI shows this with an asterisk. This diff also refuses to run the script, and warns the user, if they try to run a script with such an argument missing.

In this image, `defaultValue` is omitted in the vis spec, the argument is marked with an asterisk, the value is missing, and the run button was pressed.
{F115650}

Test Plan: Open a script, remove the `defaultValue` field from one of its variables in the vis spec, and then try to run it with the arg missing. It should refuse, and tell you to specify the arg.

Reviewers: michelle, vihang, #engineering

Reviewed By: michelle, #engineering

Subscribers: nserrino

JIRA Issues: PC-548

Differential Revision: https://phab.corp.pixielabs.ai/D7536

GitOrigin-RevId: 0d9679bafe0c3f250a8e78af86a45d0f1f474977