Add pprof handlers to all of our go servers
Summary:
This makes it easier to grab a profile on demand if there are any
problems with the service.
Note, if running `go tool pprof` on a port forwarded service, use `https+insecure`
as the scheme to ignore the TLS cert.

Test Plan:
Deploy pixie cloud with skaffold. Portforward one of the services.
Run `go tool pprof`, ensure that we can inspect profiles.

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D7538

GitOrigin-RevId: c8db56d7dcb0afe0164bebf2180c52432ed4972b