PP-2445: Verify in the distributed splitter that pem plans can execute on pems, and kelvin plans can execute on Kelvins
Summary: Depends on D7557. This diff adds the execution of a rule to the distributed splitter. After the splitter breaks up the plan, it will verify that the PEM section of the plan doesn't have any Kelvin-only UDFs, and vice versa. If it does, it will return a compiler error so that we stop silently failing in these cases.

Test Plan: added a unit test.

Reviewers: philkuz, zasgar

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D7558

GitOrigin-RevId: c79be183787c071a666a81a26b5c0e0c51e3ff61