Add user and org info to kratos config
Summary: Kratos needs to store the user and org info (shadowing them). This will mimic the structure that we use for auth0.

Test Plan: tested on a kratos instance and was able to load and store these values. Updating these on a test instance is pretty easy - you don't need a migration. The data is stored in a json string which makes this easy.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7563

GitOrigin-RevId: f3f83cde5a3d5f0f2bca479595a74b14570ffd48