PP-2486: Propagate correct errors in query broker execution
Summary:
When debugging the metadata UDF seg fault, it became evident that even if there is an error on the agent side, the query broker will always respond with an error that says that the client (aka UI or CLI) closed the stream. This was because of the following sequence of action that resulted in the actual error being silenced:
1. Error on the agent side triggers the signal to close the client stream
2. Closing the client stream
3. Returning an error that the client stream has been closed, even though it was only closed as the result of a non-client related error.
This diff ensures that the errors are actually propagated correctly and not silenced.

Test Plan: edited tests

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2486

Differential Revision: https://phab.corp.pixielabs.ai/D7579

GitOrigin-RevId: a86cc746e874a694287a47259c112aac30ff4658