PP-2486: Fix metadata UDF seg fault
Summary: Large clusters are more likely to have lots of k8s metadata which eventually gets expired. We were noticing that large clusters had queries that would never complete. I determined that this was due to a seg fault in Kelvin (problem 1) that was then handled weirdly by the query broker (problem 2). This diff fixes problem 1, the seg faulting in metadata UDFs causing a Kelvin crash when a pod that we didn't have metadata for was queries. Problem 2 will be fixed in a subsequent diff.

Test Plan: added unit tests for missing/invalid inputs.

Reviewers: michelle, vihang, vjain, zasgar, #engineering

Reviewed By: michelle, vihang, #engineering

JIRA Issues: PP-2486

Differential Revision: https://phab.corp.pixielabs.ai/D7578

GitOrigin-RevId: 0228789fd2c44629fc64e676303994f882c574d6