Add AKS to CLI allowlist checks
Summary:
AKS clusters work with pixie, so we should add it to the cli allowlist. checking for aks clusters is more annoying than gke/eks clusters, since they dont have a special k8s version we can just check.
instead we have to take the same route that we do with minikube, where we check the azure CLI to see if the cluster is in its list of clusters.

Test Plan: created aks cluster and made sure it allowed me to deploy without warnings

Reviewers: zasgar, oazizi, vihang, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7585

GitOrigin-RevId: 97ea84fccfd0f681e79a5d303a72121ed1010ec8