Drop already applied updates correctly
Summary:
This check wasn't handling the fact that we might see the same update
multiple times. As a result we were requesting missing updates with equal from
and to values and holding on to that connection for 20 minutes until it timed
out. This should get rid of such spurious requests.

Test Plan:
The async nature of this and double delivery of stan is hard to
reproduce in a test environment. I think it's doable but will require some
rewriting of the existing test. Will be added in a followup.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7589

GitOrigin-RevId: 56f52f4538563619458f6ac3d7b1f611f6924f5d