PP-2445: Schedule Kelvin-only UDFs on Kelvin nodes
Summary:
Depends on D7558. If we detect a UDF can only run on Kelvin, we should consider that the same as a blocking node and split the plan there (with that UDF's operator as part of the Kelvin plan).
If a PEM-only UDF or a Kelvin-only UDF ends up being scheduled on the wrong type of node, we will give a compiler error (same as D7558), this is just an optimization to reduce burden on the user.
Next up: Tagging various UDFs as Kelvin or PEM only, and refactoringi/removing the old metadata detection logic in the distributed splitter.

Test Plan: added a unit test

Reviewers: philkuz, jamesbartlett, zasgar

Reviewed By: philkuz, zasgar

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D7590

GitOrigin-RevId: 846b826e930e2c5bb6249398824d0195a6ae8256