CLI: Add wrapper for flag set class so that we can have required PxL script arguments without default values
Summary:
We want to be able to have some pxl script args as "required" args (which have no default value provided). For example, the value for "node" in px/node, there is no reasonable default value that we can provide for that, since we don't know the names of the users' nodes. However, the CLI relies on the golang flag package FlagSet class to parse script args. Unfortunately this CLI parsing flag class does not have support for required args with no default value. In other words, it requires you provide a default value and it will use that value if no other value is provided. As a result, I have added a wrapper class that tracks which args are required and which have default values. It uses the flag.FlagSet class underneath to simplify the logic as much as possible, while keeping track of where we should throw the user an error that a given script arg is required for the script to run.

Next up will be integrating this class into the rest of the CLI, replacing current users of flag.FlagSet.

Test Plan: added unit tests

Reviewers: zasgar, michelle, vjain, vihang

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D7607

GitOrigin-RevId: 1ec5490ffd090e02e74637261a40fa3721c4d4f5