BPF: Avoid ifdef, which gets preprocessed away
Summary:
This was the source of a bug on OpenShift. In short, the task struct overrides were not taking effect because the preprocessor had already removed the relevant code.

General rule is that preprocessor conditionals are not allowed in our BCC code, because of the way we preprocess.

Test Plan: Testing this is hard, because it requires compiling on one kernel and running on a different one. So this is done manually.

Reviewers: yzhao, jps

Reviewed By: jps

Differential Revision: https://phab.corp.pixielabs.ai/D7613

GitOrigin-RevId: 70264969416322bffed18aa1e9957c6b367dc4eb