Fix cgroup path for openshift clusters
Summary:
we had a bug where we were not getting any PIDs in openshift clusters. it turns out that's because the cgroup path is actually slightly different ("crio" instead of "docker").
to determine whether the cgroup path should contain crio vs docker, we can look at the container ID which begins with either "docker://" or "cri-o://". this information can then be passed to the PEM and used to get the right cgroup path.

Test Plan: tried on openshift cluster

Reviewers: oazizi, zasgar, vihang

Reviewed By: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D7640

GitOrigin-RevId: 33dceb5d50e7cf7dd589343e0f170d2c9cf06b80