Implement a test for the Stirling perf profiler feature.
Summary: Implement a test for the Stirling perf profiler feature. The test is based on a toy app. that computes some Fibonacci numbers. Because the computation terminates in a predictable number of iterations, we can predict where this app will spend its time. In more detail, we compute `F52` and `F27` and expect that the ratio of time spent between these is about 2:1. This toy app. was already in a previous revision, *this* revision provides a test case with this expectation.

Test Plan: This is a test. We rely on upstream libraries to provide test assertions, etc.

Reviewers: oazizi, yzhao

Reviewed By: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D7527

GitOrigin-RevId: dd128433dd49e116ebc823c69c7c6252f073f00f