PX-88 Add crds to helm chart
Summary:
This is part of the effort to allow users to deploy Pixie without the update job. This is so customers (such as customer), can look at the YAMLs and know exactly what is getting deployed to their cluster.
I'm tackling the non-bootstrap deploy for Helm charts first. currently, the etcd and nats operators are responsible for installing the etcd/nats CRDs. however, this does not work with Helm, as it expects the CRDs to be deployed first.
this diff adds the crds to our generated helm charts. these crds are copied from the helmchart definitions for both the etcd/nats operators.

Test Plan: built an rc, checked the generated helm chart

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7650

GitOrigin-RevId: 54338be2bfe6c9121bd4bf409822bc07acc6cc1a