PX-88 Update helm charts to include all Vizier YAMLs
Summary:
This is part of the effort to allow users to deploy Pixie without the update job. This is so customers (such as customer), can look at the YAMLs and know exactly what is getting deployed to their cluster.
I'm tackling the non-bootstrap deploy for Helm charts first, then will handle `extract-yamls` and `px deploy` next.

this diff updates `generate-yamls` so that it generates templated YAMLs for all vizier yamls. this required some changes to the template code to better handle CRDs.
it also required an update to the etcd-operator deployment, so that it waits for the certprovisioner job to actually provision the etcd certs.

Test Plan: ran `px generate-yamls` to generate the YAMLs, and put them in a helm chart. deployed the helm chart with/without etcdOperator, and made sure Pixie became healthy.

Reviewers: vihang, zasgar, vjain

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7652

GitOrigin-RevId: 14de64a38b368f64fb0823fd57283f9d149b59ea